Use lambda for useJsonTypeName

+++ samples/client/others/rust/hyper/oneOf/src/models/pizza.rs
@@ -13,7 +13,7 @@ use crate::models;
 #[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
 pub struct Pizza {
     #[serde(rename = "pizzaSize", skip_serializing_if = "Option::is_none")]
-    pub pizza_size: Option<f64>,
+    pub pizza_size: Option<f32>,
     /// Hyperlink reference
     #[serde(rename = "href", skip_serializing_if = "Option::is_none")]
     pub href: Option<String>,
